authorization implemented

+++ src/api/notes/handler.js
@@ -11,8 +11,14 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this.validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this.service.addNote({ title, body, tags });
+    const noteId = await this.service.addNote({
+      title,
+      body,
+      tags,
+      credentialId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -26,7 +32,8 @@ class NotesHandler {
   }
 
   async getNotesHandler(request, h) {
-    const notes = await this.service.getNotes();
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this.service.getNotes(credentialId);
     return h.response({
       status: 'success',
       data: {
@@ -37,7 +44,11 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this.service.verifyNoteOwner(id, credentialId);
     const note = await this.service.getNoteById(id);
+
     return h.response({
       status: 'success',
       data: {
@@ -49,7 +60,9 @@ class NotesHandler {
   async putNoteByIdhandler(request, h) {
     this.validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this.service.verifyNoteOwner(id, credentialId);
     await this.service.editNoteById(id, request.payload);
 
     return h.response({
@@ -60,6 +73,9 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this.service.verifyNoteOwner(id, credentialId);
     await this.service.deleteNoteById(id);
     return h.response({
       status: 'success',
